Update Diffusive Mixing Tests
